ChatGPT-4: Initial commit
Prompt: I want to build a Google Chrome extension that works with manifest.json v3. I want the chrome extension to be called Google Maps Coordinate Copier. I want the chrome extension to scan the current webpage, find any links to Google Maps, and then find the latitude and longitude that it links to. When a user clicks the extension's icon in the top right of the browser, I want a popup to appear.

If there is a Google Maps link on the page, I want there to be 3 lines in that popup:
1. In bold I want it to say "Latitude:" and then the latitude found
2. In bold I want it to say "Longitude:" and then the longitude found
3. I want a button that says "Copy Coordinates", that when clicked copies the coordinates to the user's clipboard. After copying, I want the button text to change to "Copied" and remain changed until the user closes and opens the popup again

If there is no Google Maps link on the page, I want the popup to display error text that says "No Google Maps link found on this page :("