ChatGPT-4: Trying to fix extension not displaying or copying coordinate values
Prompt: Ok. When running now, there are no errors displayed in chrome://extensions nor in the developer tools console of the extension's popup. However, the extension is not working as expected.

When I click on the extension's icon, the popup opens, and it is displaying first the message: "No Google Maps link found on this page :(", then "Latitude:" in bold, then "Longitude:" in bold, then the "Copy Coordinates" button. There are no actual latitude and longitude values displayed, nor does clicking the button copy anything to the clipboard.

I believe that part of the problem is that the links to Google maps on the pages that I am testing go to maps.google.com not google.com/maps